backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
django-storages==1.14.2
boto3==1.34.34
python-jose==3.3.0
Pillow==9.5.0
dj-database-url==2.1.0
setuptools==68.2.2
wheel==0.41.2 